[FEAT] HC-74 폴더 구조 develop 브랜치 구조로 변경
