--- src/main/java/com/example/coconote/api/block/entity/Method.java
@@ -1,6 +1,0 @@
-package com.example.coconote.api.block.entity;
-
-public enum Method {
-    create, update, delete
-//    실제 html 태그 : h, p, ol, ul , li
-}
