--- src/main/java/com/example/coconote/api/block/entity/Type.java
@@ -1,6 +1,0 @@
-package com.example.coconote.api.block.entity;
-
-public enum Type {
-    heading, paragraph, orderedList, bulletList, listItem
-//    실제 html 태그 : h, p, ol, ul , li
-}
