+++ src/main/java/com/example/coconote/api/canvas/block/entity/Method.java
@@ -1,0 +1,6 @@
+package com.example.coconote.api.canvas.block.entity;
+
+public enum Method {
+    create, update, delete
+//    실제 html 태그 : h, p, ol, ul , li
+}
