+++ src/main/java/com/example/coconote/api/canvas/canvas/dto/response/CreateCanvasResDto.java
@@ -1,0 +1,28 @@
+package com.example.coconote.api.canvas.canvas.dto.response;
+
+import com.example.coconote.api.canvas.canvas.entity.Canvas;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CreateCanvasResDto {
+    private Long canvasId;
+    private String title;
+    private Long parentCanvasId;
+
+    private Long channelId;
+
+    public static CreateCanvasResDto fromEntity(Canvas canvas) {
+        return CreateCanvasResDto.builder()
+                .canvasId(canvas.getId())
+                .title(canvas.getTitle())
+                .parentCanvasId(canvas.getParentCanvas() == null ? null : canvas.getParentCanvas().getId())
+                .channelId(canvas.getChannel() == null ? null : canvas.getChannel().getChannelId())
+                .build();
+    }
+}
