--- src/main/java/com/example/coconote/api/canvas/dto/request/ChatMessage.java
@@ -1,15 +1,0 @@
-package com.example.coconote.api.canvas.dto.request;
-
-import lombok.Data;
-
-@Data
-public class ChatMessage {
-    // 메시지 타입 : 입장, 채팅
-    public enum MessageType {
-        ENTER, TALK, CANVAS
-    }
-    private MessageType type; // 메시지 타입
-    private String roomId; // 방번호
-    private String sender; // 메시지 보낸사람
-    private String message; // 메시지
-}
