--- src/main/java/com/example/coconote/api/canvas/dto/request/ChatRoom.java
@@ -1,28 +1,0 @@
-package com.example.coconote.api.canvas.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ChatRoom implements Serializable {
-
-    private static final long serialVersionUID = 6494678977089006639L;
-
-    private String roomId;
-    private String name;
-
-    public static ChatRoom create(String name) {
-        ChatRoom chatRoom = new ChatRoom();
-        chatRoom.roomId = UUID.randomUUID().toString();
-        chatRoom.name = name;
-        return chatRoom;
-    }
-}
