--- src/main/java/com/example/coconote/api/canvas/dto/response/CanvasListResDto.java
@@ -1,18 +1,0 @@
-package com.example.coconote.api.canvas.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class CanvasListResDto {
-    private Long id;
-    private String title;
-//    private List<CanvasListResDto> childCanvas;
-}
